Update MessageDecoder interface and improve logging
Modified the MessageDecoder interface to now take in a WebSocketClient as an additional parameter. Made changes to existing methods to accommodate this change. Also improved the logging in various places: changed the log level for nil messages from 'Warn' to 'Debug', added logging when sending messages, and added an info log when the sending loop finishes.